Fix state of latest edition for /business-finance-support-finder
This path represents the old custom finder for business_support documents
that were managed by publisher.  This was replaced by a standard finder
published by specialist-publisher, and the documents are now also managed
by specialist-publisher.  The routes are different for the old and new
finders and this was handled by a redirect defined in router-data.  We
want to unpublish the current route with a new redirect content item so
we can remove the router-data entry.  One blocker to this is that the old
finder is tagged to a bunch of stuff (browse pages, taxons, etc...) that
mean it shows up in search, but the new one isn't.  When we tried to load
the old finder in content-tagger to see what it was tagged to it would
not load. We can see some of the tags via its content-store representation,
but this only shows us live tags, content-tagger will also show us draft
tags which is why we want to use it.  The reason it won't load is that
even though there is a live version of /business-finance-support-finder
in content-store, the latest edition in publishing-api (which is what
content-tagger users to map from a path to a content-id to load the item)
is marked as unpublished, and content-tagger won't load unpublished items.

We did some investigation and it's not clear at all why the latest edition
is unpublished, but there's still a live representation in content-store.
The easiest thing for us to do is to reset the state and remove the
unpublishing object for this latest edition.  We could do this through the
api but that would ultimately cause the routes to be republished from
publishing-api and that would clobber the redirect put in place by
router-data.  Ultimately getting rid of that router-data route is our goal
but we're not ready to do it before we've tagged the new finder and
unpublished the old finder as a redirect.

This PR therefore runs a migration that fixes the state of the latest
edition, and will break if the world is not as we expect.