Make business-finance-support-finder migration a no-op
The original implementation of this migration would raise exceptions if
the state of the world was not as it expected.  Now that we've merged,
deployed, and run this migration it's possible that a dev with a
mismatched set of database backups could try to run this migration and
it would break giving them no clear instructions on how to proceed.  To
avoid this we remove the implementation of the migration.  It's safe
to do so because data replication is the real way to make sure your db
is up to date.