Add migration to fix taxons visible_to_department_editors
This value should be `true` for all published editions, since this
will now control whether taxons are visible in the taxonomy sidebar.

We are doing work in ContentTagger to set this value to true once
a taxon is published so that the state is maintained properly going
forward.

See https://github.com/alphagov/content-tagger/pull/516