Use standard buildProject in Jenkinsfile
Using the standard buildProject defined in our groovy library means we get
consistent behaviour across all applications, and allows us to improve our
CI process in one place.

We do need to take advantage of some of the customisation hooks to keep
the build behaviour the same though.

1. Add extra parameters to the build for configuring which branches to use
   in the the content-store and publishing end-to-end test jobs
2. use beforeTest to set some extra env vars for coverage and pact tests
3. set publishingE2ETests to true so we always run the publishing
   end-to-end tests (our custom jenkinsfile always ran these but
   buildProject only does it if you tell it to)
4. use afterTest to publish the output of rcov
5. use afterTest to run pact verification and publish the output
6. use afterTest to run pact verification against content-store

The test task in our custom script was to run `rspec` whereas the default
in buildProject is to run `rake default`.  We could use overrideTestTask
here, but it turns out that our `rake default` behaviour is to run rspec
anyway so while it would be marginally faster to run `rspec` directly
rather than go via `rake`, consistency wins out.

There was also some difference between how our custom script ran the
publishing end-to-end tests and how buildProject does it, but as far
as I can tell the eventual outcome is the same.  The major difference
was that we would always run it whereas buildProject is conditional, but
all the parameters come from the same place (or resolve to the same
values).