Add 'with_draft' option to '/lookup-by-base-path' endpoint
During the development of the Topic Taxonomy in Content Tagger, a lot of
taxons are still in their draft states. Recently, we've had the
requirement to lookup taxons by their base path to provide an
informative error message when attempting to create a taxon with a
conflicting base path.

By introducing the 'with_draft' parameter, we can optionally include
draft states and content from both the live and draft content stores.