Document lookup-by-base-path when base path has two publishing states
Earlier in the year, work was paused on the /v2/lookup-by-base-path
request [1]. This sceanario has been added to the lookups spec and we
record what the response is with and without the recently added
`with_drafts` paramater.

This parameter is optional, and with its default value, it does not
change the response from the original. When the request is made with the
`with_drafts` parameter as true, then draft documents take precedence
over published.

[1] https://github.com/alphagov/publishing-api/pull/822