Copy publisher_first_published_at from edition to edition
This fixes some data accuracy issues that were encountered with the
backfill script in: https://github.com/alphagov/publishing-api/pull/1042

The logic when writing this initially was that we should set the
`publisher_first_published_at` field if a user specifies it and whilst
this logic is sound it is not compatible with `first_published_at` field
where once a publisher has set this once the next value will be copied
each time. This led to backfilling the dates to not being compatible
with current behaviour.