Log invalid publishing queue messages
Log messages which don't match the relevant notification schema to help
us find inconsistencies between the publishing API and the content
schemas.

Apps which consume these messages rely on the schemas to know what
fields to expect.

Just log to a file for now. We may log to Sentry once we have an idea of
the error volume and we have fixed the most common errors.