Update to the latest working version of Oj 2.
Using Oj 2.18 fixes a bug when parsing the JSON schema: publishing-api
uses `oj_mimic_json`, which makes Oj behave like the json gem. In older
versions, this mode caused the wrong exception to be thrown when parsing
invalid JSON. See
https://github.com/ruby-json-schema/json-schema/issues/305

Fixing that bug will let us use govuk-content-schema-test-helpers. It
depends on json-schema, which uses Oj for JSON parsing when Oj is
available.

This does not upgrade Oj to the very latest version:

- Oj 3 causes a lot of test failures, so may need some work to upgrade.
- Oj 2.18.5 exists on RubyGems, but it is not documented in the Oj
  changelogs, and breaks the error handling for invalid JSON in the
ContentItemsController.