Include the root_taxon link type in link expansion
The root_taxon link type is used to connect taxons to the homepage,
which is considered the root of the taxonomy.

Including the homepage in the expanded representation of the links for
taxons and for content tagged to taxons means that it's possible to
get the entire ancestry for the taxons involved just from the expanded
links.