Add represent_downstream task based on publishing date
Add a new `represent_downstream:published_between` rake task which
republishes anything which was last updated in the given time range.

It only republishes documents in the "published" state, so it ignores
drafts and documents which have been superseded.

This will make it easier to republish documents which are missing from
the search index because they were published in a certain time window.
For example, if we have had to restore a search index from backup we
will lose recent publishing events.