Add a rake task to generate a CSV of live documents
This takes a list of publishing apps and extracts all live editions
published by those apps.  The editions are written to a CSV including
URL, title, document type, and first published at.

This report is intended to give a list of documents to the content
team to help them work out what content might need uprating.