Send payload version without querying on the content_id
In 11178765670842cca8124a79dc8509dd6b322518 we made to change to
take the payload_version from the maximum ID of all events
associated with the content_id. Previously the payload_version
was coming from the ID of the event which triggered the main
publishing event.

This PR slightly reverts the behaviour to instead use the
payload_version taken from the maximum ID of any event rather
than an event associated with the content_id.

This fixes a problem we have where publishing a document which
has links to other documents would cause those linked documents
to be republished, but using their last event ID as the payload
version rather than the event ID of the original publish, meaning
that it will arrive at the content store with a payload_version
that isn't greater than the version stored in the content store
and therefore end with a 409 conflict error.

An alternative to this PR would be to create a new event for each
linked document when publishing a document.