Add rake task to populate primary publishing organisation
The primary publishing organisation is a link type that is set
by some publishing apps, but not all of them.

We are going to update all apps to set this link, as it
helps us to group together content so we can manage and
report on it better.

This rake task will allow us to update existing content by
publishing app or document type.

For example:

- Mainstream content has a primary organisation of GDS,
  because GDS are always involved in publishing it
- HMRC manuals are always published by HMRC
- Specialist publisher content is associated with different
  organisations, depending on the document type, eg AAIB reports belong to AAIB