Task to queue messages for all content
Example usage: queue:requeue_all_the_things[bulk.data-warehouse]

Currently in content performance manager, the way we initially
populate our content data is very different to how we incrementally
update it as content changes. This is because the API returns us
data in a different format to the message queue.

In particular, it doesn't send us links or the details hash,
so we have to make additional calls to the content store to
fully populate the content. These calls are architecturally
undesirable because the content store is supposed to be used
for serving the frontend of the site.

We are thinking that it would be simpler
if we could drive everything from the message queue (this
is the solution previously adopted by search). To do this
we need some way of sending all the live content items through
the system on demand. This is expected to be run only as a developer
task, so we would pause the normal consumer in content performance
manager, run this rake task, and wait for everything to be
processed.