Rake task to create drafts of special routes
Special routes tend not to have a publishing application and are
normally rather simple. This is a rake task to create them. This generic
task was done instead of creating a specific task for "/design-system"
special route.

Hopefully the generic nature of this pays off in the future - but it
wasn't much more complex than a dedicated task.