Ignore errors from sending published content to replace more recent draft
It's a common and valid occurance in the Publishing API for a piece of content to be re-rendered and try to replace a newer draft. This is a safeguard that stops that happening, and not an error. The other types of `DownstreamInvalidStateError` are ones we want to know about, so instead we should use a new subclass here and ignore that instead.