Fix code to log invalid links in patch links
We shouldn't be allowing publishing apps to send any old rubbish to the publishing API.
This code is supposed to log when links are invalid, but it didn't actually do anything.
For now, we just want to log to sentry. Once sentry is reporting no new errors coming in,
we can make this a 400 error (similar to put content).

That said, the request can currently accept links for content that doesn't
exist; i.e. the publishing app calls patch links *before* creating a draft.
I'm not sure if we actually need to support this use case?

As long as we do support this there will always been an edge case where invalid links
can sneak into the publishing API. But checking what we can is better than not doing
the check at all.