Decrease Sidekiq concurrency from 10 to 6
Currently the load on the PostgreSQL primary machine is often in the
mid 20s. This is a bit too high for a machine with now 8 cores, and it
looks like the culprit is the Publishing API Sidekiq workers.

Drop the connections back down to 6, to hopefully reduce the load on
the machine.