Get represent_downstream to work more on current editions
Previously the document_type, rendering_app and publishing_app tasks
would consider any edition ever meeting this criteria. So they would
act on editions that didn't meet the criteria, because another edition
in the past did.

While this might be useful, it's not very intuitive, so have it only
consider current editions.