Rake task to use high priority queue for represent_downstream
This task allows us to resend items to the high priority queue.

I went for a syntax of `represent_downstream:high_priority:content_id`
as we couldn't use a second option in the content_id method and we don't
have a precedent in this app for using env vars for rake tasks. If
this needs extending to other represent_downstream tasks then it
probably makes sense to have some sort of global option for them, but I
don't really anticipate it as we really shouldn't be putting anything
more than the odd thing through the high priority queue.

This feature has been added in response to it being needed during a
period of publishing sensitive documents and it being done via the
console. This will instead allow it to be run through Jenkins.