Create a multi-column index on edition id and content_store
There are quite a few examples of a query being run which uses
both of these fields, for example:

SELECT  "editions"."id"
FROM "editions"
INNER JOIN "documents" ON "documents"."id" = "editions"."document_id"
WHERE "editions"."content_store" = 'live' AND ("editions"."id" > 1308589)
ORDER BY "editions"."id" ASC LIMIT 1000

I'm unable to find in the code where the query is coming from, but
it is happening. It currently takes about 50 seconds to run cold
because although there is an index on `content_store` it's not
being used, instead the primary key index is used with a filter
on the `content_store` field. By adding a multi-column index on
both of those fields, it gets utilised in this query taking the
speed down to under a second.