Allow schema to determine whether base_path is required
This removes the storing of which schemas allow pathless content on an
edition. This was unnecessary since this information is already held in
schemas.

This also allows base_path of nil to be part of a payload, even though
this has been permitted at a schema level this was causing Publishing
API errors because of the use of `has_key?(:base_path)` which returns
true even when base_path is nil.

The change here is to resolve issues which have meant most Roles that
are stored in Whitehall have not been published. These have been failing
to publish because:

- Whitehall payload includes a nil base_path [1]
- The EMPTY_BASE_PATH_FORMATS constant had been set with role
  document_types rather than the schema name of role [2]
- We don't use EMPTY_BASE_PATH_FORMATS in a consistent way [3] [4]

[1]: https://github.com/alphagov/whitehall/blob/f723899dd3e1b8e403e255751b16204efeef9d9e/app/presenters/publishing_api/role_presenter.rb#L53
[2]: https://github.com/alphagov/publishing-api/blob/76679ce5fc1d300408a80a9bc1a14e30b6f24151/app/models/edition.rb#L35-L52
[3]: https://github.com/alphagov/publishing-api/blob/8a450d41bab60f9c0459421815af9a49b7dbe1c6/app/commands/v2/put_content.rb#L140
[4]: https://github.com/alphagov/publishing-api/blob/8a450d41bab60f9c0459421815af9a49b7dbe1c6/app/models/edition.rb#L105