Validation error when schema does not exist
Prior to this change we would not validate incoming requests when the
schema did not exist and would alert a warning. This behaviour has led
to us storing bits of schema data hardcoded in the Publishing API.

This changes the behaviour so that whenever we have a schema we don't
know we fail validation. This isn't a situation that occurs in
day-to-day use of the Publishing API (I can't find a single instance of
the warning in sentry) and is likely a holdover from the transition to
validating schemas.