Update tests to validate against actual schemas
Now that the Publishing API requires schemas to exist this updates all
the test cases where we have relied on not validating a schema we don't
know. It also updates the Edition factory which meant tests that rely on
the default value for that also needed to be changed.

This was quite a useful exercise as it revealed a number of bugs where
our tests was passing the Publishing API invalid data that a user could
never do with a known schema. Most interesting of these was the
bulk_publishing option for PutContent which could not be executed if a
schema was validated.

This has highlighted we have a pretty grim coupling issue here between
what is tested here and govuk_content_schemas. I couldn't think of a
good way to solve it here - where we could still get the benefits of
more accurate schemas - but someone is going to have a bad day if the
generic schema is updated :-/ We presumably could use a helper from
GovukContentSchemas that can mock and register a schema to a basic
specification.