Remove test for sending invalid JSON to controller
With updated oj this throws an exception as Rails tries to decode it.
So really all the test is doing is checking that Rails tries to decode
the request body - which we know it will.

Even without the exception, this is caught by the PutContentValidator
returning a 422 due to missing fields.  So there's no point in having
the test.