Expand subfields with the same nesting as the input
It was suprising to me to learn that when specifying specific
subfields to expand during link expansion, only the last part of
key is used in the output. This means for a travel advice with a
details field of `country`, it would appear in the link expansion
output under the root (not within a `details` hash).

This makes it harder to migrate formats that are currently
presenting the entire details hash as they expect the fields to be
inside a details hash.