Make healthcheck tests less brittle
I've encountered these tests [failing a][1] [few times][2] because the test environment
had a long sidekiq latency.

I've amended the tests so we can assert that the right tests are present, and
the right structure of data is returned.

The internals of the healthcheck is tested in the govuk_app_config gem's tests
so I'm not too worried about asserting the responses here.

[1]: https://ci.integration.publishing.service.gov.uk/job/publishing-api/job/deployed-to-production/842/console
[2]: https://ci.integration.publishing.service.gov.uk/job/publishing-api/job/deployed-to-production/836/console