Don't include public_updated_at in organisation links
When organisations are published in Whitehall, they are always
updated as a major update. This means that the public_updated_at
field gets updated every time meaning dependency resolution and
link expansion is always happening.

I don't think this field is being used anywhere so I think we can
probably drop it.