Query class to get links for edition ids
This query class takes an input of edition ids and looks up the links
for them and returns the result in a hash.

E.g. for an array of ids [1, 2, 3] we'd get a response of

```
{
  1 => { "link_type" => [content_id, content_id],
         "different_link_type" => [content_id] },
  2 => { "link_type" => [content_id] },
  3 => { "link_type" => [content_id] },
}
```

The intention is to use this class as part of traversing through
editions so that these can be returned as part of a HTTP request to
return editions.