Add a post pagination step to keyset pagination query
Keyset pagination works by doing a single focused query, for things that
are more complex than we can access the data with a second query once we
know which are ids are required.

This sets up a post_pagination step which can be used for the later
query. This is used to access the links of an edition at the point of
populating these results.

It also changes some of the API. I've dropped usage of the call method.
That method name seems to make sense when a class has a single method to
do it's stuff but this class has multiple entry points and classes that
interact it with need to know about the class API.

There is some cleanup about how we use strings and symbols to make them
consistent in object expectation and utilising hash fetch methods rather
than relying on OR conditions.