Refactor KeysetPagination for post pagination data
This changes the KeysetPagination query to look up post pagination data
as part of loading the results. It adds in a step to call the
post_pagination method on the client.

It now has two variables memo'd, both results and
ordered_initial_results, as we need to frequently access the
ordered_initial_results and we don't want to have to run the
post_pagination query multiple times.

It renames call being the key method of this class which doesn't really
make sense since there are multiple methods a class needs to know to
interface with this class.