Include edition links when presenting to message queue
This fixes an inconsistency where we were presenting the full expanded
links (with edition and link set links) to the message queue but were
only sending link set links as unexpanded links.

This caused an issue for Content Publisher that mostly uses edition
links. Since an organisation is tagged to a Content Publisher document
via an edition link, this information was not making it to Email Alert
API. Because of this content published via Content Publisher was not
making it to organisation based email lists.

It was a bit weird to use the Queries::LinksForEditionIds class for this
but I couldn't find a different class that provided this functionality
already, it could probably just do with a refactored interface.