Fix unsupported argument type in rake task
Passing the args from a rake task directly in to Document#find_by!
results in an "Unsupported argument type" error. I think this is because
rake args are actually an object of type Rake::TaskArguments> rather
than a hash that Document#find_by! expects.

Converting the arguments into a hash by calling #to_hash on them seems
to fix this.