Create a republish endpoint
This endpoint is to fulfill a need Content Publisher has in the ability
to transition an unpublished item back to being published. At the time
of writing this is specifically for the ability to revert a withdrawal.

This provides an endpoint called republish that can be used by any live
edition and will update the content store and message queue with the
updated document using the existing update_type of "republish".

Editions that are published won't have their state changed, just the
publishing_request_id updated. Whereas unpublished editions will have
their state set back to published, their unpublishing object will also
be removed as it's presence can cause strange things to occur.

Initially this endpoint was going to be part of the publish endpoint to
work in a similar way to the unpublish one where it can be called
multiple times to change states. However this proved to just make
something complex more complicated and raised questions about how to
indicate whether you're dealing with draft or live editions.