Default to running the master branch of GDS API Adapter pacts
This changes this from the latest published pacts by default. It's
actually unusual that we'd want to do that as that just reflects
whatever the last published branch of GDS API Adapters is. Instead it
makes more sense to run against the master branch by default, which I
believe is what was originally intended by latest.