Add a new endpoint to unreserve a path
This adds a new command to unreserve (destroy) a PathReservation,
provided the reservation exists and matches the given publishing app.

As we migrate formats from Whitehall to Content Publisher, users
sometimes need to revert to Whitehall if certain features aren't
available. Because Whitehall assumes it is authoritative for all paths
in a format, we need to be able to cleanup the Publishing API in order
for the user to re-create their content in Whitehall and re-use a path.