Adds some documenation around publishing intents
Up to now there is little to no documentation around publishing intents
in the publising-api. We used publishing intents in content-publisher
when working on the scheduling workflow and had to trawl through code to
find out how they worked and how to interact with the endpoint.

This adds some guidance around using publishing intents.

Trello:
https://trello.com/c/fXtvPpIO/696-create-publish-intent