Set change note timestamp to public_updated_at when updated
When a PUT content request is made and the payload has a change note along with
a `public_updated_at` timestamp we now set the change note's `public_timestamp`
to the `public_updated_at` value. Previously we just used the current time for
`public_timestamp`, however this is incorrect as it could cause the change
note's publicly displayed timestamp to be different to the document's publicly
displayed updated at timestamp. For example, if a document was published today
but was backdated to yesterday in its publishing app then the publicly
displayed updated at timestamp would be yesterday, but the change note would be
dated as today.

If the payload does not have a `public_updated_at` timestamp then we will
default the change note's `public_timestamp` to now, as per the previous
behaviour.