Set change note timestamp to updated_at on publish
When an edition is published we now set the change note's `public_timestamp` to
the edition's `public_updated_at` value. A change note is only available when a
major update is published and its timestamp should be consistent with
`public_updated_at` which is displayed on the frontend and is also only updated
when a major updated is published.