Simplify Publishing API edition dates
A long time ago (https://github.com/alphagov/publishing-api/pull/1005)
there was an idea to split the dates stored on an edition so it was
clear whether the dates were set by the publishing application or
automatically by the Publishing API.

Work has not been able to resume on this since then.

Picking this up now I felt it would be least disruptive to not change
the means to generate the dates and instead focus on storing the
timestamps Publishing API performed actions at.

This has culminated in dropping all the timestamp fields that store when
a publisher set something and renaming the temporary fields to represent
just a timestamp Publishing API did something at.