public_updated_at should fallback to first_published_at over now
It's possible, but undesirable, for content to be published as a minor
change on it's first publishing. This can make the timestamps strange
for an edition as first_published_at and public_updated_at can end up
referring to different times. This updates the code so that
first_published_at timestamp is used as a fallback rather than now so
that these will be more consistent.