Remove edition diff code
A long time ago efforts were made by the then Content History team to store
the difference between changes of content to make way for allowing
published content to be patched. This was built on the expectation that
we would intend to slowly expose more parts of GOV.UK history (as
stored by the Publishing API) and thus had reasons to keep a cleaner
history of editions.

2 years later the Publishing API is no nearer to allowing the patching
of published content and the drive to expose a public history of the
Publishing API is no longer present. Thus the work that was
done to store this data is likely not presenting us with any value and
is instead incurring quite a large cost given the additional size this
adds to the database (extra 10GB and growing).

The simplest thing to do is to remove this, we can always dig out this
code again if it's something that's later useful.