Increase the number of Sidekiq workers
The queues are often very long so increasing the number of workers might
help to push content downstream quicker.

The last time we tried this the load on our PostgreSQL machine went up
too much, but since then we've upgraded the machine from 8 cores to 16
cores so it might cope better this time.