Add source stats to dependency resolution worker
If the worker is provided with information about the source of the
dependency resolution (command, document type, fields) it will send
those to statsd.

This will allow us to build a picture about what is triggering
dependency resolution and give us pointers on what we should work on to
improve it in the future.