Only record dependency resolution fields if not a new document
Currently we record the fields as a source of dependency resolution if
the document is new. This means we always get fields such as
`content_id`, `document_type`, etc showing as a source of dependency
resolution. This is not useful to us because really we only want to know
which fields of a document are changing that causes dependency
resolution (because publishing a new document will always trigger
dependency resolution and there's nothing for us to improve there).