Custom expansion fields can handle multiple links
This replaces the `find_custom_expansion_fields` method to be replaced
by two methods that are used to return fields used for expanded links.

Looking up a document type without a link will return all the possible
fields this document type could have in link expansion scenarios - thus
a collation of all the varieties. We look up the field for a link when
we are deciding whether or not to perform dependency resolution. To do
this accurately we need to know every possible field that may be
represented in link data for this document_type, the only way to do this
is to combine all the data from the various link types.

Looking up document_type and link will try find fields set for that
specific scenario, if not it'll see if there are fields for that
document type, before falling back to the default fields. This resolves
an issue where previously looking up with document_type and link_type
would first match on a document_type and no link.

This change is done for the introduction of new custom fields that use
link type and document type when there is already specific fields for
that document type (step_by_step_nav).