Add DRAFT_ONLY_FIELDS to expansion rules
This specifies fields that are only represented as part of the
presentation of a drafts link expansion. This is for data that only has
relevance to the draft stack.

The auth_bypass_ids field only has relevance to the draft stack as we
only require auth to access content there. In setting this up I debated
for a while whether it mattered whether this field was sent to the live
stack and came to the conclusion that it probably doesn't matter but
since it's of no use there it's cleaner for it not to exist. As the code
to remove this is pretty simple it seemed fine to do so.

This updates the API to the expand_fields methods which requires a few
modifications to methods that call it.