Resolve inconsistent describe/context lay out of spec
This changes this spec so most scenarios are within describe blocks
rather than context ones, which resolves falling into multiple levels of
context blocks.

This is something of a plaster over wider problems in this file where
a lot of the scenarios are somewhat hard to follow. In doing this I
didn't review them thoroughly and just looked to merely resolve the
initial problem of inconsistent describe/context fields.