Update PutContent command to update edition with auth_bypass_ids
Updates the edition with the passed in auth_bypass_ids,
these can either come from the nested access_limited hash or by the
auth_bypass_ids param passed at the root level.

These two variations need to exist as not every publishing app will be
updated in this piece of work to use the updated format of not having
auth_bypass_ids in the access limit hash.