Update EditionPresenter to present root auth_bypass_ids
This code allows the EditionPresenter to present the auth_bypass_ids to
the content-store as a root edition value instead of being contained within
an access limit.

If there is auth_bypass_ids present on the edition then this value will
be amended to the access_limited hash for now to preserve the format
content-store expects to receive.

auth_bypass_ids are not being sent to the message queue as only live editions
should be enqueded here so the value should always be an empty array.